Fixed the memory leak in the UI manager

+++ src/Controller/Layer.hpp
@@ -17,7 +17,7 @@ class Layer
     /**
      * @brief Updates a Layer
      */
-    virtual void Update(double frameRate) = 0;
+    virtual void Update(float frameRate) = 0;
     /**
      * @brief Draw a Layer
      */
